backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
kraken==4.2.2
pytesseract==0.3.10
pillow==10.1.0
python-magic==0.4.27
scikit-learn==1.3.2
spacy==3.7.2